Create component InputContainer as common factor of input styles
